// src/screens/PaymentPage/PaySection/YandexPay/yapay.types.ts
type Listener<T> = (event: T) => void;
type UnsubscribeCallback = () => void;

interface TypedObject<T> {
  type: T;
}

/**
 * Общие типы.
 */

/**
 * Должно быть больше 0 и не содержать больше двух знаков после запятой.
 * Например: 1.12, 5.1, 10.
 */
export type Price = string;
// ...
export interface OrderTotal {
  /**
   * Текст для поля total.
   */
  label?: string;
  amount: Price;
}

export interface OrderItem {
  label: string;
  amount: Price;
}

export interface Order {
  /**
   * ID заказа на стороне продавца.
   */
  id: string;
  total: OrderTotal;
  items?: OrderItem[];
}
// ...
export enum PaymentEventType {
  Ready = 'ready',
  Abort = 'abort',
  Error = 'error',
  Process = 'process',
  Change = 'change',
}
// ...
export enum CompleteReason {
  Success = 'success',
  Error = 'error',
  Close = 'close',
  AuthRedirect = 'auth-redirect',
}

export interface CompleteMessage extends TypedObject<MessageType.Complete> {
  reason: CompleteReason;
  errors?: any;
}

export type Message = PaymentMessage | CompleteMessage;

/**
 * Настройки платежной кнопки.
 */

export type ButtonParent = HTMLElement | ShadowRoot;

export enum ButtonType {
  Simple = 'SIMPLE',
  Pay = 'PAY',
}

export enum ButtonTheme {
  White = 'WHITE',
  WhiteOutlined = 'WHITE-OUTLINED',
  Black = 'BLACK',
}

export enum ButtonWidth {
  Auto = 'AUTO',
  Max = 'MAX',
}

export interface ButtonOptions {
  /**
   * @default ButtonType.SIMPLE
   */
  type?: ButtonType;

  /**
   * @default ButtonTheme.RED
   */
  theme?: ButtonTheme;

  /**
   * @default ButtonWidth.AUTO
   */
  width?: ButtonWidth;
}

/**
 * События платежной кнопки.
 */

export enum ButtonEventType {
  Click = 'CLICK',
}

/**
 * Конструктор платежной кнопки.
 */

export declare class Button {
  constructor(options: ButtonOptions);

  static create(options: ButtonOptions): Button;

  /**
   * Добавить кнопку в DOM-дерева.
   */
  mount(parent: ButtonParent): void;

  /**
   * Удалить кнопку из DOM-дерева.
   */
  unmount(): void;

  /**
   * Удаляет все слушатели с кнопки и кнопку из DOM-дерева.
   */
  destroy(): void;

  /**
   * Слушатель на клик.
   * Используйте этот параметр для запуска процесса оплаты на клик по кнопке.
   */
  on(type: ButtonEventType.Click, listener: Listener<MouseEvent>): void;
}